Update README.md
The Language Trellis is a current project by Imperial undergraduates Sid Diamond, Henry Hodges and Aurelius Caesar, being supervised by Dr Heather Graven. It is an exploration into the limitations of Natural Language Processing algorithms. 

The working title is in reference to Nick Mulvey lyric: 
"And all that we ever do is learn the laws of the trellis
To better understand the rose."